feat, moved RocksDB implementation to its own project
